apps/api/pyproject.toml
[project]
name = "aiscreener-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "PyJWT>=2.8.0",
    "bcrypt>=4.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "redis>=5.0.0",
    "celery[redis]>=5.4.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-socketio>=5.11.0",
    "httpx>=0.27.0",
    "pyotp>=2.9.0",
    "qrcode[pil]>=7.4.2",
    "cryptography>=42.0.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
